chapter13/gastrodb/pyproject.toml
[project]
name = "gastrodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "bs4>=0.0.2",
    "langchain>=0.3.7",
    "pandas>=2.2.3",
    "pyarrow>=18.0.0",
    "python-dotenv>=1.0.1",
    "s3fs>=2024.10.0",
    "streamlit-chat>=0.1.1",
    "streamlit>=1.40.1",
    "typer>=0.13.0",
    "weaviate-client>=4.9.3",
    "langchain-openai>=0.2.8",
    "langchain-weaviate>=0.0.3",
    "ipykernel>=6.29.5",
]


[tool.uv]
package = true

[project.scripts]
gastrodb = 'gastrodb.cli:app'

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]


